apps/api/pyproject.toml
# Phase 0 stub — dependencies from CLAUDE.md §🏗 기술 스택
# Not all packages are installed yet; add as each phase implements its stage.

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project]
name = "verum-api"
version = "0.0.0"
requires-python = ">=3.13"
dependencies = [
    # Phase 0
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.7",
    # Phase 0 DB
    "sqlalchemy[asyncio]>=2.0",
    "alembic>=1.13",
    "asyncpg>=0.29",
    "pgvector>=0.3",
    # Phase 1 — ANALYZE
    "tree-sitter>=0.23",
    "tree-sitter-typescript>=0.23",
    "typer>=0.12",
    # Phase 2 — INFER + HARVEST
    "anthropic>=0.49",
    "httpx>=0.27",
    "trafilatura>=1.12",
    # Phase 2.5 — Auth
    "python-jose[cryptography]>=3.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5",
    "respx>=0.21",
    "ruff>=0.4",
    "pylint>=3",
    "bandit>=1.7",
    "mypy>=1.10",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.coverage.run]
omit = ["src/loop/analyze/cli.py", "src/loop/infer/cli.py", "src/loop/harvest/cli.py"]

[tool.coverage.report]
exclude_lines = ["if __name__ == .__main__.:"]

[tool.ruff]
line-length = 100

[tool.mypy]
python_version = "3.13"
strict = true
